% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipDB.R
\docType{class}
\name{chipDB-class}
\alias{chipDB-class}
\alias{.chipDB}
\alias{chipDB}
\title{chipDB-class}
\usage{
chipDB(quiet = TRUE)
}
\arguments{
\item{quiet}{A logical object, default TRUE}
}
\value{
An object of the chipDB class
}
\description{
Unified access to local ISB ChIP-seq databases
}
