% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipDB.R
\docType{methods}
\name{getGenomeName,chipDB-method}
\alias{getGenomeName,chipDB-method}
\alias{getGenomeName}
\title{get the name of the genome currently in use: TODO mapping to hg19 could be added}
\usage{
\S4method{getGenomeName}{chipDB}(obj)
}
\arguments{
\item{obj}{An object of class chipDB}
}
\description{
get the name of the genome currently in use: TODO mapping to hg19 could be added
}
