% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipDB.R
\docType{methods}
\name{getHits,chipDB-method}
\alias{getHits,chipDB-method}
\alias{getHits}
\title{get the ChiP-seq hits in the specified region, optionally limited to a specified TF}
\usage{
\S4method{getHits}{chipDB}(obj, chrom, start, end, TF = NA_character_)
}
\arguments{
\item{obj}{An object of class chipDB}

\item{chrom}{A character object}

\item{start}{numeric}

\item{end}{numeric}

\item{TF}{A character object}
}
\description{
get the ChiP-seq hits in the specified region, optionally limited to a specified TF
}
